Add Fate die support
This is kinda rough, but should work fine. A physical Fate die is a d6
containing two - signs, two + signs, and two blank sides. This is
emulated using -1 for the - sign, +1 for the + sign, and a 0 for the
blank side. The result is that we can still add them up (so 10dF work),
subtract them, keep highest etcetera.

Also added a FateLiteral class that's optional, it's just there to make
the end result (when looking at `result.expr`) a bit nicer.